split projects and workspaces tests
